+++ pkg/appmanager/appManager.go
@@ -1050,9 +1050,6 @@ func (appMgr *Manager) getQueueLength() int {
 	for _, ns := range appMgr.GetAllWatchedNamespaces() {
 		services, err := appMgr.kubeClient.CoreV1().Services(ns).List(context.TODO(), metav1.ListOptions{})
 		qLen += len(services.Items)
-		if len(services.Items) != 0 {
-			qLen++
-		}
 		if err != nil {
 			log.Errorf("[CORE] Failed getting Services from watched namespace : %v.", err)
 			qLen = appMgr.vsQueue.Len()
@@ -1066,7 +1063,6 @@ func (appMgr *Manager) processNextVirtualServer() bool {
 	if !appMgr.steadyState && appMgr.processedItems == 0 {
 		appMgr.queueLen = appMgr.getQueueLength()
 	}
-
 	if quit {
 		// The controller is shutting down.
 		return false
@@ -1074,18 +1070,26 @@ func (appMgr *Manager) processNextVirtualServer() bool {
 
 	defer appMgr.vsQueue.Done(key)
 	skey := key.(serviceQueueKey)
-	if !appMgr.steadyState && (skey.ResourceKind != Services && skey.ResourceKind != Configmaps && skey.ResourceKind != Routes) {
+	if !appMgr.steadyState {
+		//If not in steady state update,delete events are added back to queue
+		//processed once in steady state
 		if skey.Operation != OprTypeCreate {
 			appMgr.vsQueue.AddRateLimited(key)
+			return true
 		}
-		appMgr.vsQueue.Forget(key)
-		return true
+		//if it is Ingress its processed with associated service, skip this key
+		if skey.ResourceKind != Services && skey.ResourceKind != Configmaps && skey.ResourceKind != Routes {
+			appMgr.vsQueue.Forget(key)
+			return true
+		}
 	}
-
 	err := appMgr.syncVirtualServer(skey)
 	if err == nil {
 		if !appMgr.steadyState {
 			appMgr.processedItems++
+			//add service to processedresource map
+			//so same service with create event will be ignored
+			appMgr.processedResources[prepareResourceKey(Services, skey.Namespace, skey.ServiceName)] = true
 		}
 		appMgr.vsQueue.Forget(key)
 		return true
@@ -1150,13 +1154,14 @@ func (appMgr *Manager) syncVirtualServer(sKey serviceQueueKey) error {
 	// Processing just one service from a namespace processes all the resources in that namespace
 	switch sKey.ResourceKind {
 	case Services:
-		rkey := Services + "_" + sKey.Namespace
+		rkey := prepareResourceKey(sKey.ResourceKind, sKey.Namespace, sKey.ResourceName)
 		if !appMgr.steadyState && sKey.Operation == OprTypeCreate {
-			if _, ok := appMgr.processedResources[rkey]; ok {
-				if !appMgr.steadyState && appMgr.processedItems >= appMgr.queueLen {
-					appMgr.deployResource()
-					appMgr.steadyState = true
-				}
+			if !appMgr.steadyState && appMgr.processedItems >= appMgr.queueLen {
+				appMgr.deployResource()
+				appMgr.steadyState = true
+			}
+			//if service key exist in processedResource map, remove this item from processedItems count.
+			if appMgr.checkProcessedResource(rkey) {
 				return nil
 			}
 			appMgr.processedResources[rkey] = true
@@ -1165,11 +1170,21 @@ func (appMgr *Manager) syncVirtualServer(sKey serviceQueueKey) error {
 		if appMgr.IsNodePort() {
 			return nil
 		}
+		resKey := prepareResourceKey(sKey.ResourceKind, sKey.Namespace, sKey.ResourceName)
+		switch sKey.Operation {
+		case OprTypeCreate:
+			if appMgr.checkProcessedResource(resKey) {
+				return nil
+			}
+		case OprTypeDelete:
+			delete(appMgr.processedResources, resKey)
+
+		}
 	case Configmaps:
 		resKey := prepareResourceKey(sKey.ResourceKind, sKey.Namespace, sKey.ResourceName)
 		switch sKey.Operation {
 		case OprTypeCreate:
-			if _, ok := appMgr.processedResources[resKey]; ok {
+			if appMgr.checkProcessedResource(resKey) {
 				return nil
 			}
 		case OprTypeDelete:
@@ -1183,9 +1198,11 @@ func (appMgr *Manager) syncVirtualServer(sKey serviceQueueKey) error {
 		// then it was handled earlier when associated service processed
 		// otherwise just mark it as processed and continue
 		case OprTypeCreate:
-			if _, ok := appMgr.processedResources[resKey]; ok {
+			if appMgr.checkProcessedResource(resKey) {
 				return nil
 			}
+			// For any resource ,the corresponding service could be considered as processed
+			appMgr.processedResources[prepareResourceKey(Services, sKey.Namespace, sKey.ServiceName)] = true
 			appMgr.processedResources[resKey] = true
 		case OprTypeDelete:
 			delete(appMgr.processedResources, resKey)
@@ -1369,7 +1386,9 @@ func (appMgr *Manager) syncConfigMaps(
 					stats.vsUpdated += 1
 				}
 				// Mark each resource as it is already processed
+				// Mark each resource and its service as it is already processed
 				// So that later the create event of the same resource will not processed, unnecessarily
+				appMgr.processedResources[prepareResourceKey(Services, sKey.Namespace, sKey.ServiceName)] = true
 				appMgr.processedResources[prepareResourceKey(Configmaps, sKey.Namespace, cm.Name)] = true
 			}
 			continue
@@ -1414,6 +1433,7 @@ func (appMgr *Manager) syncConfigMaps(
 				}
 				// Mark each resource as it is already processed
 				// So that later the create event of the same resource will not processed, unnecessarily
+				appMgr.processedResources[prepareResourceKey(Services, sKey.Namespace, sKey.ServiceName)] = true
 				appMgr.processedResources[prepareResourceKey(Configmaps, sKey.Namespace, cm.Name)] = true
 			}
 		}
@@ -1494,6 +1514,12 @@ func (appMgr *Manager) syncIngresses(
 			if ing.ObjectMeta.Namespace != sKey.Namespace {
 				continue
 			}
+			if !appMgr.steadyState && sKey.Operation == OprTypeCreate {
+				//if ingress is processed previously.skip processing on create event
+				if _, ok := appMgr.processedResources[prepareResourceKey(Ingresses, sKey.Namespace, ing.Name)]; ok {
+					continue
+				}
+			}
 			if ok := appMgr.checkV1Beta1SingleServivceIngress(ing); !ok {
 				continue
 			}
@@ -1520,8 +1546,9 @@ func (appMgr *Manager) syncIngresses(
 				_, ingFound, _ := appInf.ingInformer.GetIndexer().GetByKey(ingKey)
 				return !ingFound
 			}
-			// Mark each resource as it is already processed
+			// Mark each resource and its service as it is already processed
 			// So that later the create event of the same resource will not processed, unnecessarily
+			appMgr.processedResources[prepareResourceKey(Services, sKey.Namespace, sKey.ServiceName)] = true
 			appMgr.processedResources[prepareResourceKey(Ingresses, sKey.Namespace, ing.Name)] = true
 			depsAdded, depsRemoved := appMgr.resources.UpdateDependencies(
 				objKey, objDeps, svcDepKey, ingressLookupFunc)
@@ -1652,6 +1679,12 @@ func (appMgr *Manager) syncIngresses(
 			if ing.ObjectMeta.Namespace != sKey.Namespace {
 				continue
 			}
+			if !appMgr.steadyState && sKey.Operation == OprTypeCreate {
+				//if ingress is processed previously.skip processing on create event
+				if _, ok := appMgr.processedResources[prepareResourceKey(Ingresses, sKey.Namespace, ing.Name)]; ok {
+					continue
+				}
+			}
 			if ok := appMgr.checkV1SingleServivceIngress(ing); !ok {
 				continue
 			}
@@ -1678,8 +1711,9 @@ func (appMgr *Manager) syncIngresses(
 				_, ingFound, _ := appInf.ingInformer.GetIndexer().GetByKey(ingKey)
 				return !ingFound
 			}
-			// Mark each resource as it is already processed
+			// Mark each resource and its associated service as it is already processed
 			// So that later the create event of the same resource will not processed, unnecessarily
+			appMgr.processedResources[prepareResourceKey(Services, sKey.Namespace, sKey.ServiceName)] = true
 			appMgr.processedResources[prepareResourceKey(Ingresses, sKey.Namespace, ing.Name)] = true
 			depsAdded, depsRemoved := appMgr.resources.UpdateDependencies(
 				objKey, objDeps, svcDepKey, ingressLookupFunc)
@@ -1852,9 +1886,10 @@ func (appMgr *Manager) syncRoutes(
 			continue
 		}
 
-		// Mark each resource as it is already processed during init Time
+		// Mark each resource  as it is already processed during init Time
 		// So that later the create event of the same resource will not processed, unnecessarily
 		appMgr.processedResources[prepareResourceKey(Routes, sKey.Namespace, route.Name)] = true
+		appMgr.processedResources[prepareResourceKey(Services, sKey.Namespace, sKey.ServiceName)] = true
 
 		key := route.Spec.Host + route.Spec.Path
 		if host, ok := routePathMap[key]; ok {
@@ -2917,7 +2952,6 @@ func (appMgr *Manager) getEndpointsForCluster(
 	// Mark each resource as it is already processed
 	// So that later the create event of the same resource will not processed, unnecessarily
 	appMgr.processedResources[prepareResourceKey(Endpoints, eps.Namespace, eps.Name)] = true
-
 	for _, subset := range eps.Subsets {
 		for _, p := range subset.Ports {
 			if portName == p.Name {
@@ -3289,4 +3323,15 @@ func getIngressV1Backend(ing *netv1.Ingress) []string {
 		svcs = append(svcs, key)
 	}
 	return svcs
+}
+
+func (appMgr *Manager) checkProcessedResource(rkey string) bool {
+	if _, ok := appMgr.processedResources[rkey]; ok {
+		//if service key exist in processedResource, remove this item from processedItems count.
+		if !appMgr.steadyState {
+			appMgr.processedItems--
+		}
+		return true
+	}
+	return false
 }